chore: move lineno converters to yaml library
